ANDKR | Code Analyze 1차 수정

+++ dailyHOTEL/app/src/main/java/com/daily/dailyhotel/screen/home/gourmet/detail/menus/GourmetMenusView.java
@@ -383,7 +383,7 @@ public class GourmetMenusView extends BaseDialogView<GourmetMenusView.OnEventLis
 
             if (timeTag != null)
             {
-                if (time == (int) timeTag)
+                if (time == timeTag)
                 {
                     childView.setSelected(true);
                 } else
